+++ src/common/authentication/IAccountSASSignatureValues.ts
@@ -1,9 +1,9 @@
-import { SasIPRange as IIPRange } from "@azure/storage-blob";
+import { SasIPRange } from "@azure/storage-blob";
 
 import {
   computeHMACSHA256,
   truncatedISO8061Date
-} from "../../common/utils/utils";
+} from "../utils/utils";
 import AccountSASPermissions from "./AccountSASPermissions";
 import AccountSASResourceTypes from "./AccountSASResourceTypes";
 import AccountSASServices from "./AccountSASServices";
@@ -84,10 +84,10 @@ export interface IAccountSASSignatureValues {
   /**
    * Optional. IP range allowed.
    *
-   * @type {IIPRange | string}
+   * @type {SasIPRange | string}
    * @memberof IAccountSASSignatureValues
    */
-  ipRange?: IIPRange | string;
+  ipRange?: SasIPRange | string;
 
   /**
    * The values that indicate the services accessible with this SAS. Please refer to {@link AccountSASServices} to
@@ -106,6 +106,15 @@ export interface IAccountSASSignatureValues {
    * @memberof IAccountSASSignatureValues
    */
   resourceTypes: AccountSASResourceTypes | string;
+
+  /**
+   * Indicates the encryption scope to use to encrypt the request contents.
+   * This field is supported with version 2020-12-06 or later.
+   *
+   * @type {string}
+   * @memberof IAccountSASSignatureValues
+   */
+   encryptionScope?: string;
 }
 
 /**
@@ -120,6 +129,77 @@ export interface IAccountSASSignatureValues {
  * @memberof IAccountSASSignatureValues
  */
 export function generateAccountSASSignature(
+  accountSASSignatureValues: IAccountSASSignatureValues,
+  accountName: string,
+  sharedKey: Buffer
+): [string, string] {
+
+  if (accountSASSignatureValues.version >= "2020-12-06") {
+    return generateAccountSASSignature20201206(
+      accountSASSignatureValues,
+      accountName,
+      sharedKey
+    );
+  }
+  else {
+    return generateAccountSASSignature20150405(
+      accountSASSignatureValues,
+      accountName,
+      sharedKey
+    );
+  }
+}
+
+function generateAccountSASSignature20201206(
+  accountSASSignatureValues: IAccountSASSignatureValues,
+  accountName: string,
+  sharedKey: Buffer
+): [string, string] {
+  const parsedPermissions = accountSASSignatureValues.permissions.toString();
+  const parsedServices = accountSASSignatureValues.services.toString();
+  const parsedResourceTypes = accountSASSignatureValues.resourceTypes.toString();
+  const parsedStartTime =
+    accountSASSignatureValues.startTime === undefined
+      ? ""
+      : typeof accountSASSignatureValues.startTime === "string"
+      ? accountSASSignatureValues.startTime
+      : truncatedISO8061Date(accountSASSignatureValues.startTime, false);
+  const parsedExpiryTime =
+    typeof accountSASSignatureValues.expiryTime === "string"
+      ? accountSASSignatureValues.expiryTime
+      : truncatedISO8061Date(accountSASSignatureValues.expiryTime, false);
+  const parsedIPRange =
+    accountSASSignatureValues.ipRange === undefined
+      ? ""
+      : typeof accountSASSignatureValues.ipRange === "string"
+      ? accountSASSignatureValues.ipRange
+      : ipRangeToString(accountSASSignatureValues.ipRange);
+  const parsedProtocol =
+    accountSASSignatureValues.protocol === undefined
+      ? ""
+      : accountSASSignatureValues.protocol;
+  const version = accountSASSignatureValues.version;
+  const encryptionScope = accountSASSignatureValues.encryptionScope;
+
+  const stringToSign = [
+    accountName,
+    parsedPermissions,
+    parsedServices,
+    parsedResourceTypes,
+    parsedStartTime,
+    parsedExpiryTime,
+    parsedIPRange,
+    parsedProtocol,
+    version,
+    encryptionScope,
+    "" // Account SAS requires an additional newline character
+  ].join("\n");
+
+  const signature: string = computeHMACSHA256(stringToSign, sharedKey);
+  return [signature, stringToSign];
+}
+
+function generateAccountSASSignature20150405(
   accountSASSignatureValues: IAccountSASSignatureValues,
   accountName: string,
   sharedKey: Buffer