+++ src/table/authentication/ITableSASSignatureValues.ts
@@ -2,8 +2,8 @@ import {
   computeHMACSHA256,
   truncatedISO8061Date
 } from "../../common/utils/utils";
-import { SASProtocol } from "./IAccountSASSignatureValues";
-import { IIPRange, ipRangeToString } from "./IIPRange";
+import { SASProtocol } from "../../common/authentication/IAccountSASSignatureValues";
+import { IIPRange, ipRangeToString } from "../../common/authentication/IIPRange";
 
 /**
  * ITableSASSignatureValues is used to help generating Table service SAS tokens for containers or tables.