+++ tests/queue/apis/queueService.test.ts
@@ -256,7 +256,7 @@ describe("QueueServiceAPIs", () => {
     await serviceClient.getStatistics()
       .catch((err) => {
         assert.strictEqual(err.statusCode, 400);
-        assert.strictEqual(err.details.Code, "InvalidQueryParameterValue");
+        assert.strictEqual(err.details.code, "InvalidQueryParameterValue");
         assert.ok(err);
       });;
   });